restructure files within folder hierarchy
