+++ src/components/admin/PublishedCheckIn.jsx
@@ -16,7 +16,7 @@ import { Link as ReactRouterLink } from 'react-router-dom';
 import { Link as ChakraLink } from '@chakra-ui/react';
 import Loading from '../shared/Loading';
 import { useMutation } from '@tanstack/react-query';
-import useAdminService from '../../services/adminService';
+import useAdminService from '../../hooks/services/useAdminService';
 import LocalStorageService from '../../util/LocalStorageService';
 
 const PublishedCheckIn = () => {
